feat: create PCITextfield to less boilerplate in the fields
